+++ packages/anatomy/src/carousel/index.ts
@@ -1,3 +1,0 @@
-import { anatomy as _ } from '@zag-js/carousel'
-
-export const anatomy = _.extendWith('control')