+++ packages/anatomy/src/checkbox/index.ts
@@ -1,3 +1,0 @@
-import { anatomy as _ } from '@zag-js/checkbox'
-
-export const anatomy = _.extendWith('group')