+++ packages/anatomy/src/date-picker/index.ts
@@ -1,3 +1,0 @@
-import { anatomy as _ } from '@zag-js/date-picker'
-
-export const anatomy = _.extendWith('view')