+++ packages/anatomy/src/segment-group/index.ts
@@ -1,3 +1,0 @@
-import { anatomy as _ } from '@zag-js/radio-group'
-
-export const anatomy = _.rename('segment-group')