+++ website/src/app/license/page.tsx
@@ -1,30 +1,40 @@
 import type { Metadata } from 'next'
+import { notFound } from 'next/navigation'
 import { Container, Stack } from 'styled-system/jsx'
 import { PageHeader } from '~/components/page-header'
 import { Prose } from '~/components/ui/prose'
 import { MDXContent } from '~/mdx-content'
 import { pages } from '.velite'
 
-const license = pages.find((page) => page.id === 'license')
-
 export default function Page() {
+  const page = getPageById('license')
+  if (!page) {
+    return notFound()
+  }
+
   return (
-    <Container py="12" maxW="7xl">
-      <Stack gap="12" alignItems="center">
+    <Container py={{ base: '16', md: '24' }} maxW="3xl">
+      <Stack gap={{ base: '16', md: '24' }} alignItems="center">
         <PageHeader
-          heading="Ark UI Plus"
-          subHeading="License"
-          description="License agreement for Ark UI Plus"
+          heading={page.title}
+          subHeading={page.subtitle ?? ''}
+          description={page.description}
         />
         <Prose>
-          <MDXContent code={license?.code ?? ''} />
+          <MDXContent code={page.code} />
         </Prose>
       </Stack>
     </Container>
   )
 }
 
-export const metadata: Metadata = {
-  title: 'Ark UI Plus - License',
-  description: 'License agreement for Ark UI Plus',
+const getPageById = (id: string) => pages.find((page) => page.id === id)
+
+export async function generateMetadata(): Promise<Metadata> {
+  const page = getPageById('license')
+
+  if (page) {
+    return { title: page.title, description: page.description }
+  }
+  return {}
 }